Renamed search.jsx to searchresult.jsx
